Create 5 rounds for the game
Create function game() and playerPlay()

The function game loops the compareWeapon for 5 times for 5 rounds. the playerWeapon is now returned inside the function since it needs to be called in every loop. The new function playerPlay() is called in game so that everytime the loop begins, the user can enter new value.